Changed minSdk and app namespace & id
