+++ examples/object_detection/android-jetpack-compose/app/src/main/java/com/google/mediapipe/examples/objectdetection/ui/theme/Color.kt
@@ -1,0 +1,12 @@
+package com.google.mediapipe.examples.objectdetection.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Turquoise = Color(0xFF009688)
+val TurquoiseGrey = Color(0xFF355E5A)
+val LightBlue = Color(0xFF00BCD4)
+
